Added droop DLVC sim, changed DLVC sim current sampling, commented param
Made an equivalent DLVC sim for the droop controller as I had previously
made for the dVOC controller. I tested the DLVC with droop control and
found it to work with faster voltage loop gains. Still seeing issues
with the dVOC controller simulation. In both these simulations I changed
the current sampling to be on the grid side and the plotted power is the
power at the point of common coupling (PCC).

Added organizational comments to the params of some simulations.